requirements.txt
anyio==3.6.1
APScheduler==3.9.1
backports.zoneinfo==0.2.1
black==22.6.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
dnspython==2.2.1
fastapi==0.78.0
h11==0.13.0
idna==3.3
msoffcrypto-tool==5.0.0
mypy-extensions==0.4.3
numpy==1.23.0
olefile==0.46
pandas==2.0.2
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
pydantic==1.9.1
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
typing-extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uvicorn==0.18.2
xlrd==2.0.1